Update to Chromium revision 0bfd25d4 (#381305) - Delete include/cef_runnable.h (issue #1336). - Build the cef_unittests target using all Chromium headers. Add a   USING_CHROMIUM_INCLUDES define and libcef_dll_wrapper_unittests   target to support this. This change avoids compile errors due to   the divergence of CEF and Chromium base/ header implementations.   The libcef_dll_wrapper sources must now compile successfully with   both CEF and Chromium base/ headers (issue #1632). - The onbeforeunload message specified via JavaScript is no longer   passed to the client (see http://crbug.com/587940).

+++ libcef/browser/osr/render_widget_host_view_osr_mac.mm
@@ -5,6 +5,7 @@
 
 #include "libcef/browser/osr/render_widget_host_view_osr.h"
 
+#include <algorithm>
 #include <limits>
 #include <utility>
 
@@ -267,7 +268,8 @@ bool CefRenderWidgetHostViewOSR::GetLineBreakIndex(
   // assume the line breaking as the next character's y offset is larger than
   // a threshold. Currently the threshold is determined as minimum y offset plus
   // 75% of maximum height.
-  const size_t loop_end_idx = std::min(bounds.size(), range.end());
+  const size_t loop_end_idx =
+      std::min(bounds.size(), static_cast<size_t>(range.end()));
   int max_height = 0;
   int min_y_offset = std::numeric_limits<int32_t>::max();
   for (size_t idx = range.start(); idx < loop_end_idx; ++idx) {
